chore: initilised xo, validate-branch-name and commit-lint to run in husky pre-commit and pre-push hooks
