--- src/types/favouriteConversionButtonType.ts
@@ -1,7 +1,0 @@
-export type favouriteConversionButtonType = {
-	theme: string;
-	fromUnitDetails: string;
-	toUnitDetails: string;
-	allFavourites: Object;
-	setAllFavourites: React.Dispatch<React.SetStateAction<Object>>;
-}
