+++ src/types/from-to-units-options-inputs.ts
@@ -13,6 +13,6 @@ export type FromToUnitsOptionsInputsType = {
 	setPreviousCategoryValue: React.Dispatch<React.SetStateAction<string>>;
 	sectionHeightSmall: string;
 	sectionHeightLarge: string;
-	allFavourites: Object;
-	setAllFavourites: React.Dispatch<React.SetStateAction<Object>>;
+	allFavourites: Record<string, unknown>;
+	setAllFavourites: React.Dispatch<React.SetStateAction<Record<string, unknown>>>;
 };
